# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/register": {
      "post": {
        "description": "Creates a pending_verification user and emails a 6-digit OTP. The code is only returned in the response when inviteToken matches a live invitation for the same address (ORG-AUTH-001).",
        "operationId": "AuthController_register",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Registered; verification pending.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUser"
                }
              }
            }
          },
          "400": {
            "description": "Request failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited: 5 attempts per 15 minutes per IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Register a person",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/verify-email": {
      "post": {
        "description": "Activates the account and issues a first session.",
        "operationId": "AuthController_verify",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Verified and signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionTokens"
                }
              }
            }
          },
          "400": {
            "description": "Request failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited: 5 attempts per 15 minutes per IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Verify an email address with the OTP",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "description": "Wrong credentials answer with one generic error whether or not the address exists, so the endpoint cannot be used to enumerate users. Five failures in 15 minutes locks the account for 30.",
        "operationId": "AuthController_login",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionTokens"
                }
              }
            }
          },
          "400": {
            "description": "Request failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited: 5 attempts per 15 minutes per IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Sign in",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "description": "Refresh tokens rotate: the presented token is invalid afterwards.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "New token pair.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionTokens"
                }
              }
            }
          },
          "400": {
            "description": "Request failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited: 5 attempts per 15 minutes per IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Exchange a refresh token for a new pair",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/forgot-password": {
      "post": {
        "description": "Always reports success, whether or not the address is registered.",
        "operationId": "AuthController_forgot",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Reset email sent if the address exists."
          },
          "400": {
            "description": "Request failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited: 5 attempts per 15 minutes per IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Start a password reset",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/reset-password": {
      "post": {
        "description": "Consumes the token and revokes every existing session for the user.",
        "operationId": "AuthController_reset",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Password replaced; all sessions revoked."
          },
          "400": {
            "description": "Request failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited: 5 attempts per 15 minutes per IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Complete a password reset",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/mfa/enable": {
      "post": {
        "description": "Not implemented yet. ORG-AUTH-003 is a P1 carried forward from Sprint 1.",
        "operationId": "AuthController_enableMfa",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Request failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to do this.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Enable TOTP multi-factor authentication",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/sessions": {
      "get": {
        "operationId": "AuthController_sessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Active sessions.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionSummary"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to do this.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List active sessions for the signed-in user",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/sessions/{sessionId}": {
      "delete": {
        "description": "Signs out a single device.",
        "operationId": "AuthController_revoke",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Session revoked."
          },
          "400": {
            "description": "Request failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to do this.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Revoke one session",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_doLogout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Signed out."
          },
          "400": {
            "description": "Request failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to do this.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Sign out of the current session",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/logout-all": {
      "post": {
        "operationId": "AuthController_doLogoutAll",
        "parameters": [],
        "responses": {
          "204": {
            "description": "All sessions revoked."
          },
          "400": {
            "description": "Request failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to do this.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Sign out of every device",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zoqo API",
    "description": "Zoqo is a self-hosted communication hub. Every response is an envelope:\n`{ \"success\": true, \"data\": ... }` on success, `{ \"success\": false, \"error\": { code, message, requestId } }`\non failure (SRS §18.2). Quote the `requestId` — also returned as the `X-Request-Id` header — when reporting a problem.\n\nAuthenticate with the access token from `POST /v1/auth/login`. Organization-scoped routes also need the\n`X-Org-Id` header naming the organization you are acting in.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token, 15 minute lifetime."
      }
    },
    "schemas": {
      "PublicUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending_verification",
              "active",
              "locked",
              "suspended"
            ]
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "status"
        ]
      },
      "ApiErrorBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR",
            "description": "Stable machine-readable code."
          },
          "message": {
            "type": "string",
            "example": "Email is required"
          },
          "requestId": {
            "type": "string",
            "example": "req_9f3c1a7b2d04",
            "description": "Also returned in the X-Request-Id header; quote it in bug reports."
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ]
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
      "SessionTokens": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT, 15 minute lifetime."
          },
          "refreshToken": {
            "type": "string",
            "description": "Opaque, 30 day lifetime, rotated on refresh."
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "SessionSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "device": {
            "type": "string",
            "example": "Firefox on Windows"
          },
          "ip": {
            "type": "string",
            "example": "203.0.113.4"
          },
          "location": {
            "type": "string",
            "example": "unknown"
          },
          "lastActiveAt": {
            "type": "string",
            "format": "date-time"
          },
          "current": {
            "type": "boolean",
            "description": "True for the session making this request."
          }
        },
        "required": [
          "id",
          "device",
          "ip",
          "location",
          "lastActiveAt",
          "current"
        ]
      },
...
    }
  }
}